move main GUI class
