+++ platform-sdk/consensus-gui/build.gradle.kts
@@ -5,3 +5,8 @@ plugins {
 }
 
 description = "Consensus Hashgraph GUI"
+
+testModuleInfo {
+    requires("org.hiero.consensus.hashgraph.impl.test.fixtures")
+    requires("com.swirlds.common.test.fixtures")
+}
