+++ platform-sdk/consensus-gui/src/test/java/org/hiero/consensus/gui/HashgraphGuiMain.java
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-package org.hiero.consensus.hashgraph.impl;
+package org.hiero.consensus.gui;
 
 import com.swirlds.common.context.PlatformContext;
 import com.swirlds.common.test.fixtures.platform.TestPlatformContextBuilder;
